+++ test/testdata/type_checking_imports2.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+import typing
+
+import type_checking_imports3
+
+assert type_checking_imports3
+
+if typing.TYPE_CHECKING:
+    from type_checking_imports3 import StrOrInt
+
+    assert StrOrInt