+++ test/testdata/type_checking_imports3.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+import typing
+
+if typing.TYPE_CHECKING:
+    StrOrInt: typing.TypeAlias = "str | int"